// helpers/schema.tsx
/**
 * This file was generated by kysely-codegen.
 * Please do not edit it manually.
 */

import type { ColumnType } from "kysely";

export type Generated<T> = T extends ColumnType<infer S, infer I, infer U>
  ? ColumnType<S, I | undefined, U>
  : ColumnType<T, T | undefined, T>;

export type Timestamp = ColumnType<Date, Date | string, Date | string>;

export type UserRole = "admin" | "user";

export interface BlogPosts {
  authorId: number;
  content: string;
  createdAt: Generated<Timestamp>;
  id: Generated<number>;
  published: Generated<boolean>;
  slug: string;
  title: string;
  updatedAt: Generated<Timestamp>;
}

export interface Comments {
  content: string;
  createdAt: Generated<Timestamp>;
  id: Generated<number>;
  postId: number;
  userId: number;
}

export interface Events {
  createdAt: Generated<Timestamp | null>;
  createdBy: number;
  description: string | null;
  eventDate: Timestamp;
  id: Generated<number>;
  title: string;
  updatedAt: Generated<Timestamp | null>;
}

export interface Likes {
  createdAt: Generated<Timestamp>;
  id: Generated<number>;
  postId: number;
  userId: number;
}

export interface LoginAttempts {
  attemptedAt: Generated<Timestamp | null>;
  email: string;
  id: Generated<number>;
  success: Generated<boolean | null>;
}

export interface OauthAccounts {
  createdAt: Generated<Timestamp | null>;
  id: Generated<number>;
  provider: string;
  providerEmail: string;
  providerUserId: string;
  updatedAt: Generated<Timestamp | null>;
  userId: number;
}

export interface OauthStates {
  codeVerifier: string;
  createdAt: Generated<Timestamp | null>;
  expiresAt: Timestamp;
  id: Generated<number>;
  provider: string;
  redirectUrl: string;
  state: string;
}

export interface Sessions {
  createdAt: Generated<Timestamp | null>;
  expiresAt: Timestamp;
  id: string;
  lastAccessed: Generated<Timestamp | null>;
  userId: number;
}

export interface UserPasswords {
  createdAt: Generated<Timestamp | null>;
  id: Generated<number>;
  passwordHash: string;
  userId: number;
}

export interface Users {
  avatarUrl: string | null;
  createdAt: Generated<Timestamp | null>;
  displayName: string;
  email: string;
  id: Generated<number>;
  role: Generated<UserRole>;
  updatedAt: Generated<Timestamp | null>;
}

export interface DB {
  blogPosts: BlogPosts;
  comments: Comments;
  events: Events;
  likes: Likes;
  loginAttempts: LoginAttempts;
  oauthAccounts: OauthAccounts;
  oauthStates: OauthStates;
  sessions: Sessions;
  userPasswords: UserPasswords;
  users: Users;
}


// Use these for zod schema enums or other purposes, since the kysely enums are just types
export const UserRoleArrayValues: [UserRole, ...UserRole[]] = ["admin","user"];
// Note: kysely maps the table/column/enum names from snake_case to camelCase and PascalCase. When running SQL statements, make sure to use snake_case, but TypeScript code should use camelCase or PascalCase.